refactor(numba): trim _detect to HAS_NUMBA; avoid float64 image copy
- _detect.py: drop the unused HAS_JAX / HAS_JAX_GPU flags. Besides being dead
  for this PR, HAS_JAX_GPU eagerly imported jax at module load whenever jax was
  installed, just to set a flag nothing reads. jax detection lands with the jax
  backend; HAS_NUMBA alone establishes the find_spec pattern.
- flatten the image without a forced float64 copy: pass masked_image through
  ascontiguousarray without dtype=, and let flatten_numba upcast the kept values
  into its float64 output. Avoids a full-image float64 temporary for non-float64
  inputs (e.g. float32 microscopy data); bit-identical for float64.

+++ src/cp_measure/core/numba/measureobjectintensity.py
@@ -86,7 +86,7 @@ def get_intensity(
 
     values, seg0, xc, yc, zc = flatten_numba(
         numpy.ascontiguousarray(masks),
-        numpy.ascontiguousarray(masked_image, dtype=numpy.float64),
+        numpy.ascontiguousarray(masked_image),
         lut,
     )
     has_objects = values.size > 0
